Signals rework

+++ src/arion.cpp
@@ -102,14 +102,14 @@ void ArionGroup::run()
         {
             auto weak_instance = *instance_it;
             std::shared_ptr<Arion> instance = weak_instance.second;
-            if (!instance->is_zombie)
+            if (!instance->is_zombie() && !instance->is_stopped())
             {
                 if (!instance->run_current())
                 {
                     if (instance->has_parent())
                     {
                         std::shared_ptr<Arion> parent = instance->get_parent();
-                        instance->is_zombie = true;
+                        instance->set_zombie();
                         parent->send_signal(instance->get_pid(), SIGCHLD);
                     }
                     else
@@ -184,6 +184,7 @@ std::shared_ptr<Arion> Arion::new_instance(std::vector<std::string> program_args
     arion->sock = SocketManager::initialize(arion);
     arion->hooks = HooksManager::initialize(arion);
     arion->threads = ThreadingManager::initialize(arion);
+    arion->signals = SignalManager::initialize(arion);
     arion->tracer = CodeTracer::initialize(arion);
     arion->init_program(prog_parser);
     return arion;
@@ -202,6 +203,7 @@ Arion::~Arion()
     this->syscalls.reset();
     this->abi.reset();
     this->tracer.reset();
+    this->signals.reset();
     this->threads.reset();
     this->hooks.reset();
     this->sock.reset();
@@ -308,7 +310,6 @@ bool Arion::run_current()
     if (!threads_count)
         return false;
     bool multi_thread = threads_count > 1;
-    this->threads->handle_signals();
     if (this->threads->is_curr_locked())
     {
         this->threads->switch_to_next_thread();
@@ -606,13 +607,37 @@ void Arion::reset_group()
     this->group.reset();
 }
 
+bool Arion::is_stopped()
+{
+    return this->stopped;
+}
+
+void Arion::set_stopped()
+{
+    this->stopped = true;
+}
+
+void Arion::set_resumed()
+{
+    this->stopped = false;
+}
+
+bool Arion::is_zombie()
+{
+    return this->zombie;
+}
+
+void Arion::set_zombie()
+{
+    this->zombie = true;
+}
+
 void Arion::send_signal(pid_t source_pid, int signo)
 {
     if (this->afl_mode &&
         std::find(this->afl_signals.begin(), this->afl_signals.end(), signo) != this->afl_signals.end())
         abort(); // Will cause a crash in AFL instance
-    std::shared_ptr<SIGNAL> sig = std::make_shared<SIGNAL>(source_pid, signo);
-    this->pending_signals.push_back(sig);
+    this->signals->handle_signal(source_pid, signo);
 }
 
 void Arion::run_gdbserver(uint32_t port)
